requirements.txt
Appium-Python-Client==2.9.0
attrs==23.1.0
certifi==2023.7.22
h11==0.14.0
idna==3.4
outcome==1.3.0.post0
PySocks==1.7.1
selenium==4.9.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==1.26.18
wsproto==1.2.0
